add function in tractor.js to iterate through nested arrays in plan and call necessary functions to populate field. debugged. call tractor function in main.js
